requirements.txt
datasets>=4.0.0
langchain>=0.3.0
sentence-transformers>=5.0.0
faiss-cpu>=1.7.0
chromadb>=1.0.0
transformers>=5.0.0
torch>=2.0.0
rouge-score>=0.1.2
nltk>=3.8
numpy>=1.24.0
accelerate>=1.0.0
